requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.1
langserve[all]==0.1.0
pypdf==4.2.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
httpx==0.27.0
langsmith==0.1.0